requirements.txt
passlib[bcrypt]
pydantic
pydantic-settings
pyotp
tortoise-orm[asyncpg]
fastapi
uvicorn[standart]
cloud-sql-python-connector[asyncpg]
loguru
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-jose
python-multipart
aiosmtplib
aerich